Fix the test toolchain_broken_symlink on Windows
`rustup` on Windows uses "directory junctions" instead of symbolic links
for `rustup toolchain link`. These are somewhere between a symbolic and
a hard link, and require fewer permissions. However, the test in
question was trying to use symbolic links. It now uses the same codepath
to create the link as `toolchain link` itself.