fix(doc): don't show the opening message when --path is used
This might seem like a travel or unnecessary thing, especially since the message
is written to `stderr`  so it won't affect piping or similar use cases.
But since `stdout` and `stderr` are ambiguous in most terminals (i.e., they're printed in the output)
this might cause some confusion and/or annoyance to some users.