Upload Windows artifacts into correct subdirectory
The Windows artifacts are uploaded to S3 as part of every build on
`master`, but due to slight differences in the upload command syntax
between Linux and Windows currently end up at the root of the upload
directory. Similar to the existing upload step for builds on `stable`,
the command has been tweaked to place the files into the `dist`
directory.
k